seperate Lobby into container/ component, add tests
