+++ app/components/Lobby/styles.js
